Dev (#7)
* feat: store both default and custom config vals (custom are overriding default ones)

* refactor: Attributes\Controller and Attributes\Route classes are readonly now

* docs: fix RouteNotFoundException phpdoc

* feature: DI container

Major update with custom DI container realization. A lot of fixes/changes are made to adapt this feature usage. Most of them are related to:
- class constructors (no more new object creating in constructors);
- new object creation (all framework classes are acquired from the container now).

Also:
- 'TemplateManager' and 'TemplateManagerInterface' classes are removed as useless;
- new exceptions related to Container (located in Exceptions\Container\);
- new 'di_container' key in the config file, to pre-set services;
- a bit different app start - now Container class instance should be created first and then we can get the Kernel class instance from it.
- a lot of small fixes/tweaks.

* chore: gitignore 'src/sync.ffs_db' file

* feat: singleton classes attribute support

Now class to become singleton only need to have a 'Singleton' attribute.
The old way with the 'getIntsance' method is still supported too (The Container class itself still uses it, in case it will be called from outside).